numcodecs_random_projection

Function project_into

pub fn project_into<T: FloatExt, S: Data<Elem = T>>(
    data: ArrayBase<S, Ix2>,
    projected: ArrayViewMut<'_, T, Ix2>,
    projection: impl Fn(usize, usize) -> T,
    normalizer: T,
) -> Result<(), RandomProjectionCodecError>

Applies random projection to the input data and outputs into the projected array.

The random projection matrix is defined by the projection function (i, j) -> P[i, j] and a globally applied normalizer factor.

Errors

Errors with

RandomProjectionCodecError::NumberOfSamplesMismatch if the input data’s number of samples doesn’t match the projected array’s number of samples

RandomProjectionCodecError::NonFiniteData if the input data or projected output contains non-finite data