numcodecs_random_projection

Function project_with_projection

pub fn project_with_projection<T: FloatExt, S: Data<Elem = T>, D: Dimension>(
    data: ArrayBase<S, D>,
    seed: u64,
    reduction: &RandomProjectionReduction,
    projection: &RandomProjectionKind,
) -> Result<Array<T, Ix2>, RandomProjectionCodecError>

Applies random projection to the input data with the given seed, reduction method, and projection kind and returns the resulting projected array.

Errors

Errors with

RandomProjectionCodecError::NonMatrixData if the input data is not a two-dimensional matrix

RandomProjectionCodecError::NonFiniteData if the input data or projected output contains non-finite data