numcodecs_random_projection

Function reconstruct_into_with_projection

pub fn reconstruct_into_with_projection<T: FloatExt, S: Data<Elem = T>, D: Dimension>(
    projected: ArrayBase<S, D>,
    reconstructed: ArrayViewMut<'_, T, D>,
    seed: u64,
    projection: &RandomProjectionKind,
) -> Result<(), RandomProjectionCodecError>

Applies the (approximate) inverse of random projection to the projected array to reconstruct the input data with the given seed and projection kind and outputs into the reconstructed array.

Errors

Errors with

RandomProjectionCodecError::NonMatrixData if the projected and reconstructed arrays are not two-dimensional matrices

RandomProjectionCodecError::NumberOfSamplesMismatch if the number of samples $N$ of the projected array don’t match the number of samples of the reconstructed array

RandomProjectionCodecError::ProjectedArrayZeroComponents if the projected array is of shape $(N, 0)$

RandomProjectionCodecError::CorruptedNumberOfComponents if the projected array’s dimensionality metadata is corrupted

RandomProjectionCodecError::NumberOfFeaturesMismatch if the reconstructed array’s dimensionality $D$ does not match the projected array’s dimensionality metadata

RandomProjectionCodecError::NonFiniteData if the projected array or the reconstructed output contains non-finite data